Moving wasm test
to not fail when running `make test`.
