+++ rego/rego_wasmtarget_test.go
@@ -10,6 +10,7 @@ package rego
 import (
 	"context"
 	"fmt"
+	"github.com/open-policy-agent/opa/metrics"
 	"math/rand"
 	"net/http"
 	"net/http/httptest"
@@ -392,5 +393,37 @@ func TestEvalWasmWithHTTPAllowNet(t *testing.T) {
 
 	if len(requests) != 1 {
 		t.Fatal("Expected server to never be called")
+	}
+}
+
+func TestRegoTargetWasmAndTargetPluginDisablesIndexingTopdownStages(t *testing.T) {
+	tp := testPlugin{}
+	RegisterPlugin("rego.target.foo", &tp)
+	t.Cleanup(resetPlugins)
+
+	for _, tgt := range []string{"wasm", "foo"} {
+		t.Run(tgt, func(t *testing.T) {
+			m := metrics.New()
+			r := New(Query("foo = 1"), Module("foo.rego", "package x"), Metrics(m), Instrument(true), Target(tgt))
+			ctx := context.Background()
+			_, err := r.Eval(ctx)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			expAbsent := []string{
+				"timer_query_compile_stage_build_comprehension_index_ns",
+				"timer_compile_stage_rebuild_comprehension_indices_ns",
+				"timer_compile_stage_rebuild_indices_ns",
+			}
+
+			all := m.All()
+
+			for _, name := range expAbsent {
+				if _, ok := all[name]; ok {
+					t.Errorf("Expected NOT to find %v but did", name)
+				}
+			}
+		})
 	}
 }
